FirstEntry
Not been using Github as frequent as I should, so this is where I start taking control over that again.

currently following a tutorial on Vue
where most my updates will be focused around learning vue, and getting a better understanding of it!
Link if you're interested in taking a look, I highly recommend it if you're looking to get into a framework!

https://www.youtube.com/watch?v=FXpIoQ_rT_c